UMG truncation warning easy fixes
#jira UE-166271
#rb yohann.dossantos

[CL 30788028 by vincent gauthier in ue5-main branch]

+++ Engine/Source/Runtime/UMG/Private/Components/Viewport.cpp
@@ -251,7 +251,7 @@ FLinearColor FUMGViewportClient::GetBackgroundColor() const
 
 float FUMGViewportClient::GetOrthoUnitsPerPixel(const FViewport* InViewport) const
 {
-	const float SizeX = InViewport->GetSizeXY().X;
+	const int32 SizeX = InViewport->GetSizeXY().X;
 
 	// 15.0f was coming from the CAMERA_ZOOM_DIV marco, seems it was chosen arbitrarily
 	return ( GetOrthoZoom() / ( SizeX * 15.f ) )/* * ComputeOrthoZoomFactor(SizeX)*/;
